requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2025.7.21
yt-dlp[webvtt]==2025.7.21
yt-dlp[websockets]==2025.7.21
yt-dlp[ffmpeg]==2025.7.21
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
pytube==15.0.0
youtube-dl==2021.12.17
scrapy==2.11.0
playwright==1.40.0
ffmpeg-python==0.2.0
aiohttp==3.9.1
cloudscraper==1.2.71
requests[socks]==2.31.0
tqdm==4.66.1
colorama==0.4.6
minio==7.2.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0 